+++ frame/api/app.py
@@ -6,8 +6,10 @@ from fastapi.middleware.gzip import GZipMiddleware
 from frame import __version__
 from frame.utils import get_logger
 from frame.models.base import SessionLocal
+from frame.constants import MODEL_RELOAD_SECONDS
 from frame.api.services import stations as station_service
 from frame.api.namespaces.stations import router as stations_router
+from frame.api.dependencies import ETAPredictor, AvailabilityPredictor
 
 logger = get_logger(__name__)
 
@@ -28,7 +30,7 @@ app.include_router(stations_router)
 
 
 @app.on_event("startup")
-@repeat_every(seconds=86400, max_repetitions=1, logger=logger)
+@repeat_every(seconds=86400, max_repetitions=None, logger=logger)
 def refresh_stations_info() -> None:
     logger.info("Refreshing stations info")
     db = SessionLocal()
@@ -37,9 +39,24 @@ def refresh_stations_info() -> None:
 
 
 @app.on_event("startup")
-@repeat_every(seconds=30, max_repetitions=1, logger=logger)
+@repeat_every(seconds=30, max_repetitions=None, logger=logger)
 def refresh_stations_status() -> None:
     logger.info("Refreshing stations status")
     db = SessionLocal()
     station_service.update_stations_status(db)
     db.close()
+
+
+@app.on_event("startup")
+@repeat_every(
+    seconds=MODEL_RELOAD_SECONDS,
+    max_repetitions=None,
+    logger=logger,
+)
+def refresh_models() -> None:
+    logger.info("Reloading models")
+    logger.info("Reloading ETA model")
+    ETAPredictor.reload()
+    logger.info("Reloading availability model")
+    AvailabilityPredictor.reload()
+    logger.info("Models reloaded")