+++ frame/api/namespaces/stations.py
@@ -4,10 +4,15 @@ from sqlalchemy.orm import Session
 from fastapi import Depends, APIRouter, HTTPException
 
 from frame.utils import get_logger
-from frame.api.dependencies import get_db
-from frame.exceptions import StationDoesNotExist
 from frame.api.schemas import stations as station_schemas
 from frame.api.services import stations as station_service
+from frame.exceptions import PredictionError, NoInfoForStation, StationDoesNotExist
+from frame.api.dependencies import (
+    ETAPredictor,
+    MLFlowPredictor,
+    AvailabilityPredictor,
+    get_db,
+)
 
 logger = get_logger(__name__)
 
@@ -48,8 +53,23 @@ def predict_for_station(
     station_id: int,
     prediction_params: station_schemas.PredictionParams,
     db: Session = Depends(get_db),
+    eta_predictor: MLFlowPredictor = Depends(lambda: ETAPredictor),
+    availability_predictor: MLFlowPredictor = Depends(lambda: AvailabilityPredictor),
 ):
     try:
-        return station_service.predict(station_id, prediction_params, db)
+        return station_service.predict(
+            station_id,
+            prediction_params,
+            db,
+            eta_predictor=eta_predictor,
+            availability_predictor=availability_predictor,
+        )
     except StationDoesNotExist:
         raise HTTPException(status_code=404, detail="Station does not exist")
+    except NoInfoForStation:
+        raise HTTPException(
+            status_code=404,
+            detail="There is no information for this station. Probably does not exist.",
+        )
+    except PredictionError:
+        raise HTTPException(status_code=503, detail="Predictor uninitialized")