+++ frame/constants.py
@@ -2,6 +2,15 @@
 
 import enum
 import pathlib
+from typing import Dict, Callable
+
+from sklearn.metrics import (
+    mean_squared_error,
+    mean_absolute_error,
+    mean_absolute_percentage_error,
+)
+
+from frame.train.metrics import fn, fp, tn, tp
 
 DEFAULT_SQLITE_LOC = pathlib.Path(__name__).parent.parent / "frame.db"
 DEFAULT_SQLITE = f"sqlite:////{DEFAULT_SQLITE_LOC.resolve()}"
@@ -20,3 +29,44 @@ MAX_OVERFLOW: int = 200
 
 STATUS_ENDPOINT = "stationStatus"
 STATIONS_ENDPOINT = "stationInformation"
+
+
+class MLFlowStage(str, enum.Enum):
+    Staging = "Staging"
+    Production = "Production"
+    Archived = "Archived"
+
+
+class FrameModels(str, enum.Enum):
+    ETA = "eta"
+    AVAILABILITY = "availability"
+
+
+MODELS_QUERIES = {FrameModels.ETA: "eta", FrameModels.AVAILABILITY: "availability"}
+
+
+class FrameMetric(str, enum.Enum):
+    MAPE = "MAPE"
+    MAE = "MAE"
+    MSE = "MSE"
+    FN = "FN"
+    FP = "FP"
+    TN = "TN"
+    TP = "TP"
+
+
+METRICS_MAPPING: Dict[FrameMetric, Callable] = {
+    FrameMetric.MAE: mean_absolute_error,
+    FrameMetric.MSE: mean_squared_error,
+    FrameMetric.MAPE: mean_absolute_percentage_error,
+    FrameMetric.FN: fn,
+    FrameMetric.FP: fp,
+    FrameMetric.TN: tn,
+    FrameMetric.TP: tp,
+}
+
+DEFAULT_TEST_SIZE: float = 0.1
+
+FALLBACK_KEY: str = "fallback"
+
+MODEL_RELOAD_SECONDS: int = 60 * 5