+++ CHANGELOG.md
@@ -5,6 +5,23 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.0.0] - 2023-02-21
+### Changed
+- Python version constrained below 3.11
+
+### Added
+- ETA model training
+- Availability model training
+- Loading models from MLFlow in API
+- Predict with loaded models in API
+- Reading dataset from MinIO through DuckDB
+
+### Removed
+- Mocked predictions in API
+
+### Fixed
+- Stations were refreshed only once
+
 ## [0.6.0] - 2022-11-17
 ### Added
 - S3 client to transparently handle minio/s3